+++ src/main/java/toilari/otlite/view/lwjgl/TextRenderer.java
@@ -60,10 +60,10 @@ public class TextRenderer {
      * @param size   fonttikoko
      * @param string piirrettävä merkkijono
      */
-    public void draw(@NonNull LWJGLCamera camera, int x, int y, float r, float g, float b, int size, @NonNull String string) {
+    public void draw(@NonNull LWJGLCamera camera, float x, float y, float r, float g, float b, int size, @NonNull String string) {
         size = Math.max(this.minFontSize, Math.min(this.maxFontSize, size));
         string = string.toUpperCase();
-        int destX = x, destY = y;
+        float destX = x, destY = y;
         for (int i = 0; i < string.length(); i++) {
             val c = string.charAt(i);
